Factor out control sys Queue tags to new file
This is done because when we add size control, there will be many
new tags.
