+++ src/ControlSystem/ControlErrors/Shape.hpp
@@ -10,6 +10,7 @@
 
 #include "ControlSystem/Protocols/ControlError.hpp"
 #include "ControlSystem/Tags.hpp"
+#include "ControlSystem/Tags/QueueTags.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "Domain/Structure/ObjectLabel.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/SpherepackIterator.hpp"
