+++ tests/Unit/ControlSystem/ControlErrors/Test_Translation.cpp
@@ -12,6 +12,7 @@
 #include "ControlSystem/DataVectorHelpers.hpp"
 #include "ControlSystem/Tags.hpp"
 #include "ControlSystem/Tags/MeasurementTimescales.hpp"
+#include "ControlSystem/Tags/QueueTags.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "Domain/FunctionsOfTime/RegisterDerivedWithCharm.hpp"
 #include "Domain/FunctionsOfTime/Tags.hpp"
