+++ tests/Unit/Helpers/ControlSystem/SystemHelpers.hpp
@@ -32,6 +32,7 @@
 #include "ControlSystem/Tags.hpp"
 #include "ControlSystem/Tags/IsActive.hpp"
 #include "ControlSystem/Tags/MeasurementTimescales.hpp"
+#include "ControlSystem/Tags/QueueTags.hpp"
 #include "ControlSystem/TimescaleTuner.hpp"
 #include "ControlSystem/UpdateControlSystem.hpp"
 #include "DataStructures/DataBox/Tag.hpp"
